+++ versionhash.go
@@ -1,0 +1,2 @@
+package main
+var GitCommitHash = "ea8453363f53fb5feaa2120776ab6bc868012698"