+++ docs/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Next
 
-* Added `%version`
+* Added `%version`, and environment variables `GONB_VERSION`, `GONB_GIT_COMMIT`.
+* Added `%help` info on missing environment variables.
 
 ## v0.10.6, 2024/10/16, Improved Docker, added `%capture`